refactor(backend): adiciona refatoração certificado
Adicionando refatoração do certificado.
