+++ projeto/src/routes/certificados.routes.js
@@ -1,41 +1,11 @@
 const { Router } = require("express");
 const {
-  findCertificadoByHash,
-} = require("../repositories/certificados.repository");
+  getCertificadoByHashController,
+} = require("../controllers/certificados.controller");
 
 const router = Router();
 
-router.get("/hash/:hash", async function (req, res) {
-  const certificadoHash = String(req.params.hash || "").trim();
-
-  if (!certificadoHash) {
-    return res.status(400).json({
-      message: "Hash do certificado obrigatório",
-    });
-  }
-
-  try {
-    const certificado = await findCertificadoByHash(certificadoHash);
-
-    if (!certificado) {
-      return res.status(404).json({
-        message: "Certificado inexistente para o hash informado",
-      });
-    }
-
-    if (certificado.indisponivel) {
-      return res.status(409).json({
-        message: certificado.motivo,
-      });
-    }
-
-    return res.status(200).json(certificado);
-  } catch (e) {
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
+router.get("/hash/:hash", getCertificadoByHashController);
 
 module.exports = router;
